+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+* Cleanup of output logical ids
+* Add parameters for VPC CIDRs
+* Upgrade CDK to 1.87.1
+
 # 1.4.0
 
 * Adding Util class